pyproject.toml
[project]
name = "petals"
version = "0.1.0"
description = "End-to-end data platform demonstration"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "boto3>=1.42.30",
    "constructs>=10.0.0",
    "pyarrow>=22.0.0",
    "pyiceberg>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
